requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
databricks-connect==17.3.3
databricks-sdk==0.77.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
idna==3.11
numpy==2.4.1
packaging==25.0
pandas==2.3.3
protobuf==6.33.4
py4j==0.10.9.9
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
zstandard==0.25.0
